requirements.txt
langchain==0.0.208
deeplake==3.6.12
openai==0.27.8
tiktoken==0.4.0
python-dotenv
black==23.7.0
isort==5.12.0
autoflake==2.2.0